Comparison of several supervised classification algorithms using an E-Mail dataset
We have a set of emails, half of which were written by one person and
the other half by another person at the same company . Our objective is
to classify the emails as written by one person or the other based only
on the text of the email.  We'll tackle this problem using Naive Bayes,
SVM and Decision Tree algorithms and explore similarities and
differences between these.